sauvegarde avant restructureation des composant et sous composant
