+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import HomePage from './components/HomePage/HomePage';
-import Quizz from './Components/Quizz'
+import Quizz from './Components/Quizz/Quizz';
 
 function App() {
   return (
